Simplifica a página inicial para um único ecrã
A home passa a mostrar apenas a proposta de valor ("O único curso do género
em Portugal / Avalia lojas. Recebe por isso.") e os três passos do processo,
ligados por uma curva ascendente com números fantasma, no design atual do
site (azul profundo, vermelho de energia). Ocupa exatamente o ecrã menos o
cabeçalho, sem scroll: o rodapé fica escondido apenas nesta página.

O restante conteúdo mudou-se para as páginas já existentes:
- Sobre: pilares "O que levas contigo", setores "O que vais avaliar" e a
  faixa de números.
- O Curso: secção "Como funciona" com os quatro passos. O currículo, o preço
  e o CTA final já existiam nesta página.

Os textos da home passam a vir das traduções (PT/EN), substituindo o espaço
de nomes `home` que já não era usado.
