+++ app/about/page.module.css
@@ -242,3 +242,203 @@
   align-items: center;
   justify-content: center;
 }
+
+/* ============================================================
+   SECÇÕES VINDAS DA ANTIGA PÁGINA INICIAL
+   ============================================================ */
+.sectionAlt {
+  padding: 96px 0;
+  background: var(--color-bg-light);
+  border-top: 1px solid var(--color-gray-light);
+}
+@media (max-width: 720px) { .sectionAlt { padding: 64px 0; } }
+
+.dot { color: var(--color-primary-red); }
+
+.head {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: var(--sp-md);
+  margin-bottom: var(--sp-2xl);
+  align-items: end;
+}
+@media (min-width: 880px) {
+  .head { grid-template-columns: 1.25fr 1fr; gap: var(--sp-4xl); }
+}
+.headSub {
+  font-size: 16px;
+  line-height: 1.7;
+  color: var(--muted);
+  max-width: 46ch;
+  margin: 0;
+  text-wrap: pretty;
+}
+.headKicker {
+  font-size: 13px;
+  font-weight: 600;
+  letter-spacing: 0.5px;
+  text-transform: uppercase;
+  color: var(--color-primary-red);
+  margin: var(--sp-sm) 0 0;
+}
+
+/* Pilares */
+.pillars {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: var(--sp-md);
+}
+@media (min-width: 860px) {
+  .pillars { grid-template-columns: repeat(3, 1fr); }
+}
+.pillar {
+  background: var(--color-white);
+  border: 1px solid var(--color-gray-light);
+  border-radius: var(--radius-lg);
+  padding: var(--sp-lg);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+  transition: all 300ms ease;
+}
+.pillar:hover {
+  transform: translateY(-4px);
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.12);
+  border-color: rgba(239, 62, 77, 0.2);
+}
+.pillarAccent {
+  background: var(--color-primary-red);
+  border-color: var(--color-primary-red);
+  box-shadow: 0 4px 12px rgba(239, 62, 77, 0.25);
+}
+.pillarAccent:hover {
+  border-color: var(--color-primary-red);
+  box-shadow: 0 8px 24px rgba(239, 62, 77, 0.3);
+}
+.pillarIcon {
+  display: block;
+  width: 32px;
+  height: 32px;
+  color: var(--color-primary-red);
+  margin-bottom: var(--sp-md);
+}
+.pillarIcon svg { width: 100%; height: 100%; }
+.pillarAccent .pillarIcon { color: var(--color-white); }
+.pillarTitle {
+  font-size: 22px;
+  font-weight: 700;
+  letter-spacing: -0.2px;
+  color: var(--color-text-primary);
+  margin: 0 0 12px;
+}
+.pillarAccent .pillarTitle { color: var(--color-white); }
+.pillarRule {
+  width: 34px;
+  height: 3px;
+  border-radius: var(--radius-pill);
+  background: var(--color-primary-red);
+  margin-bottom: var(--sp-sm);
+}
+.pillarAccent .pillarRule { background: var(--color-white); }
+.pillarDesc {
+  font-size: 15px;
+  line-height: 1.6;
+  color: var(--muted);
+  margin: 0;
+  text-wrap: pretty;
+}
+.pillarAccent .pillarDesc { color: rgba(255, 255, 255, 0.92); }
+
+/* Setores */
+.cats {
+  display: grid;
+  grid-template-columns: 1fr;
+  gap: var(--sp-md);
+}
+@media (min-width: 640px) {
+  .cats { grid-template-columns: repeat(2, 1fr); }
+}
+@media (min-width: 1000px) {
+  .cats { grid-template-columns: repeat(3, 1fr); }
+}
+.cat {
+  background: var(--color-white);
+  border: 1px solid var(--color-gray-light);
+  border-top: 3px solid var(--color-primary-red);
+  border-radius: var(--radius-lg);
+  padding: var(--sp-md);
+  display: flex;
+  align-items: center;
+  gap: var(--sp-sm);
+  transition: all 300ms ease;
+}
+.cat:hover {
+  transform: translateY(-4px);
+  box-shadow: 0 12px 32px rgba(0, 0, 0, 0.1);
+}
+.catIcon {
+  width: 26px;
+  height: 26px;
+  color: var(--color-primary-red);
+  flex-shrink: 0;
+}
+.catBody { flex: 1; min-width: 0; }
+.catName {
+  font-size: 20px;
+  font-weight: 700;
+  color: var(--color-primary-blue);
+  margin: 0 0 4px;
+}
+.catPay {
+  font-size: 14px;
+  color: var(--muted);
+  margin: 0;
+}
+.catPay strong {
+  color: var(--color-success-green);
+  font-weight: 700;
+  font-size: 16px;
+}
+
+/* Faixa de números */
+.strip {
+  background: var(--color-bg-light);
+  border-top: 1px solid var(--color-gray-light);
+  border-bottom: 1px solid var(--color-gray-light);
+}
+.stripGrid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+}
+@media (min-width: 860px) {
+  .stripGrid { grid-template-columns: repeat(4, 1fr); }
+}
+.stripItem {
+  padding: var(--sp-xl) var(--sp-md);
+  border-right: 1px solid var(--color-gray-light);
+  border-bottom: 1px solid var(--color-gray-light);
+}
+.stripItem:nth-child(2n) { border-right: none; }
+@media (min-width: 860px) {
+  .stripItem { border-bottom: none; }
+  .stripItem:nth-child(2n) { border-right: 1px solid var(--color-gray-light); }
+  .stripItem:last-child { border-right: none; }
+}
+.stripNum {
+  font-size: clamp(32px, 3.2vw, 44px);
+  font-weight: 700;
+  letter-spacing: -0.5px;
+  line-height: 1;
+  color: var(--color-primary-blue);
+  margin-bottom: 12px;
+}
+.stripNum em {
+  font-style: normal;
+  color: var(--color-primary-red);
+}
+.stripLabel {
+  font-size: 13px;
+  color: var(--muted);
+  line-height: 1.6;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  font-weight: 600;
+}
