+++ app/components/AppShell.tsx
@@ -5,6 +5,7 @@
 "use client";
 
 import Link from "next/link";
+import { usePathname } from "next/navigation";
 import { LanguageProvider } from "@/app/context/LanguageContext";
 import HeaderActions from "./HeaderActions";
 import LanguageSwitcher from "./LanguageSwitcher";
@@ -12,6 +13,11 @@ import TopNav from "./TopNav";
 import Footer from "./Footer";
 
 export default function AppShell({ children }: { children: React.ReactNode }) {
+  const pathname = usePathname();
+  // A página inicial ocupa exatamente um ecrã: sem rodapé não há nada para
+  // percorrer abaixo do processo. As restantes páginas mantêm o rodapé.
+  const isHome = pathname === "/";
+
   return (
     <LanguageProvider>
       <div className="mx-auto flex min-h-screen w-full flex-col bg-transparent">
@@ -51,9 +57,12 @@ export default function AppShell({ children }: { children: React.ReactNode }) {
           </div>
         </header>
 
-        <main className="flex-1 px-0 pb-0">{children}</main>
+        {/* Na página inicial o conteúdo estica-se para ocupar o ecrã inteiro. */}
+        <main className={`flex-1 px-0 pb-0 ${isHome ? "flex flex-col" : ""}`}>
+          {children}
+        </main>
 
-        <Footer />
+        {!isHome ? <Footer /> : null}
       </div>
     </LanguageProvider>
   );
